FaceID/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.70.1
frozenlist==1.3.1
fsspec==2022.10.0
future==0.18.2
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
h11==0.14.0
hnswlib==0.6.2
idna==3.4
importlib-metadata==5.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
nest-asyncio==1.5.4
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.6.0.66
orjson==3.8.0
packaging==21.3
Pillow==9.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pyDeprecate==0.3.1
pyparsing==3.0.9
python-multipart==0.0.5
pytorch-lightning==1.4.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.12.1
torchmetrics==0.10.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.16.0
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.10.0
bcolz
prettytable
menpo
mxnet